requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
reportlab==4.0.7
openpyxl==3.1.2
python-dateutil==2.8.2
pillow==10.1.0
django-cors-headers==4.3.0 
django-filter==23.1
python-dotenv==1.0.1
dj-database-url==2.1.0
